Refacto with namespaces
