+++ demos/multigrid/samurai_new/utils.cpp
@@ -1,0 +1,34 @@
+#pragma once
+#include <samurai/algorithm.hpp>
+// #include <samurai/field.hpp>
+// #include <samurai/mr/mesh.hpp>
+// #include <samurai/mr/adapt.hpp>
+
+template<class Mesh>
+void print_mesh(Mesh& mesh)
+{
+    std::cout << mesh << std::endl;
+    samurai::for_each_cell(mesh, [](const auto& cell)
+    {
+        std::cout << cell.level << " " << cell.center(0) << std::endl; 
+    });
+}
+
+
+template<class Field>
+bool check_nan_or_inf(const Field& f)
+{
+    std::size_t n = f.mesh().nb_cells();
+    bool is_nan_or_inf = false;
+    for (std::size_t i = 0; i<n; ++i)
+    {
+        double value = f.array().data()[i];
+        if (std::isnan(value) || std::isinf(value))
+        {
+            is_nan_or_inf = true;
+            break;
+        }
+    }
+    return !is_nan_or_inf;
+}
+
